Consolidate tests to one file
